Buscar Viaje finalizado

+++ database/migrations/2021_05_06_041301_create_usuarios_table.php
@@ -20,7 +20,7 @@ class CreateUsuariosTable extends Migration
             $table->integer('dni');
             $table->string('email')->unique();
             $table->string('contraseña');
-            $table->integer('tarjeta')->nullable();
+            $table->bigInteger('tarjeta')->nullable();
             $table->string('fechaVenc')->nullable();
             $table->integer('codigo')->nullable();
             $table->integer('id_membresia')->default(2);
